requirements.txt
streamlit
google-generativeai
python-dotenv
beautifulsoup4
Pillow